Support for storing and loading wrenches
Add support for storing and loading subsymbolic data about wrenches
(forces & torques) acting on objects. This works exactly like the
mechanism for loading and storing /tf.

Right now, wrench data is always defined in the local coordinate
system of the object they act on and cannot be transformed by KnowRob.